check file size

+++ ExportData/index.js
@@ -5,6 +5,7 @@ const path = require('path');
 const fs = require('fs');
 
 var database;
+var count;
 
 createFolder();
 dbInit();
@@ -42,7 +43,6 @@ function dbInit()
             console.log("DB-Error: " + err);
             return;
         }
-        console.log("Connected to database!");
         getData();
     });
 }
@@ -54,10 +54,12 @@ function getData()
         if (err) {
             console.error(err);
         } else if (result.length >= 1) {
+            count = 0;
             for(var i in result)
             {
                 readData(result[i]);
             }
+            console.log(`${count} out of ${result.length} entries copied`);
             finish();
         }
     });
@@ -73,19 +75,20 @@ function readData(data)
     let dest_label = path.join(env.DATAFOLDER, par[3], par[4]);
     let loc_label = path.join(env.GETFOLDER, par[3], par[4]);
 
-    writeData(dest_component, loc_component, dest_label, loc_label)
+    writeData(dest_component, loc_component, dest_label, loc_label, data)
 }
 
-function writeData(dest_component, loc_component, dest_label, loc_label)
+function writeData(dest_component, loc_component, dest_label, loc_label, data)
 {
+    if( (fs.statSync(loc_component).size < 1000) || (fs.statSync(loc_label).size < 1000)) return; // Delete out of DB?
     if(fs.existsSync(dest_component)) fs.unlinkSync(dest_component);
     if(fs.existsSync(dest_label)) fs.unlinkSync(dest_label);
     fs.copyFileSync(loc_component, dest_component);
     fs.copyFileSync(loc_label, dest_label);
+    count++;
 }
 
 function finish()
 {
-    console.log("done");
     database.end();
 }
